pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptlang"
version = "0.1.0"
description = "PromptLang Compiler Platform - Transform Human Input → PromptLang IR → Optimized IR → Model Dialect → Contract Enforced Output"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "PromptLang Team"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "jsonschema>=4.20.0",
    "structlog>=24.1.0",
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "PyGithub>=2.1.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "tiktoken>=0.5.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[project.scripts]
promptlang = "promptlang.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["promptlang*", "schemas*", "webapp*"]
exclude = ["docker*", "tests*", "*.tests", "*.tests.*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
